se agrego el formato en quetzales en todo el modulo de compras; se agrego el encabezado del detalle de compras; se agrego por defecto el nit y direccion en clientes; se coloco opcional agregar telefono en clientes; se cambio el tipo de dato en las migraciones en el area de compras referente a los float.

+++ database/migrations/2015_10_09_222628_create_Compras_table.php
@@ -15,10 +15,10 @@ class CreateComprasTable extends Migration
         Schema::create('Compras', function(Blueprint $table) {
             $table->increments('id');
             $table->date('fecha');
-            $table->float('subTotal');
-            $table->float('descuento');
-            $table->float('total');
-		    $table->integer('proveedores_id')->unsigned();
+            $table->decimal('subTotal', 10, 2);
+            $table->decimal('descuento', 10, 2);
+		    $table->decimal('total', 10, 2);
+            $table->integer('proveedores_id')->unsigned();
             $table->foreign('proveedores_id')->references('id')->on('Proveedores');
             $table->timestamps();
         });
